web/requirements.txt
FastAPI==0.115.2
sqlmodel==0.0.22
fastapi-cli==0.0.5
pydantic[email]==2.9.2
asyncpg==0.30.0
aiosqlite==0.20.0
greenlet==3.1.1
PyJWT==2.10.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
circomlibpy==1.0.0